% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_aggregate.R
\name{condense_dt}
\alias{condense_dt}
\title{Condense time series in a \code{data.frame} to a lower frequency}
\usage{
condense_dt(
  dates_df,
  to_period = c("weekly", "monthly", "quarterly", "annual"),
  aggregate_by = c("last", "sum", "avg", "median")
)
}
\arguments{
\item{dates_df}{\code{data.frame} containing a date column and values}

\item{to_period}{string indicating the frequency to convert data to}

\item{aggregate_by}{string indicating how to aggregate data}
}
\value{
\code{data.frame}
}
\description{
Condense time series in a \code{data.frame} to a lower frequency
}
