% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_aggregate.R
\name{expand_dt}
\alias{expand_dt}
\title{Expand time series in a data.frame from low to high frequency}
\usage{
expand_dt(dates_df, to_period = c("weekly", "monthly", "quarterly", "annual"))
}
\arguments{
\item{dates_df}{\code{data.frame} containing a date column and values}

\item{to_period}{string indicating the periodicity to convert data to}
}
\value{
data.frame
}
\description{
Expand time series in a data.frame from low to high frequency
}
