% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{get_date_col}
\alias{get_date_col}
\title{find and rename the date column}
\usage{
get_date_col(dt_data)
}
\arguments{
\item{dt_data}{data.frame containing a column of class \code{Date}}
}
\value{
data.frame with column of class \code{Date} renamed as "date" or NULL
}
\description{
find and rename the date column
}
