% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{get_date_vector}
\alias{get_date_vector}
\title{get a vector of dates from \code{data.frame}, \code{zoo}, \code{xts}, or date named vectors}
\usage{
get_date_vector(dt_data)
}
\arguments{
\item{dt_data}{object containing date data}
}
\value{
a vector of class \code{Date}
}
\description{
get a vector of dates from \code{data.frame}, \code{zoo}, \code{xts}, or date named vectors
}
