% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{get_eom_dates}
\alias{get_eom_dates}
\title{Get the end of month date}
\usage{
get_eom_dates(dates)
}
\arguments{
\item{dates}{vector of objects of class \code{Date}, \code{POSIXct}, or \code{character} with in year, month, day format}
}
\value{
vector of dates of the same size of argument \code{dates} but at month end
}
\description{
Get the end of month date
}
\examples{
get_eom_dates("2016-01-10")

}
