% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{is_weekend}
\alias{is_weekend}
\title{Determine if a date is a weekend}
\usage{
is_weekend(dates, weekend_days = c("Saturday", "Sunday"), abbreviate = FALSE)
}
\arguments{
\item{dates}{vector of class \code{Date}}

\item{weekend_days}{string indicating the days that are classified as weekend. Default to Saturday and Sunday}

\item{abbreviate}{logical indicating whether the character vector of weekend days is abbreviated or not}
}
\value{
vector of class \code{logical} indicating weekend days
}
\description{
Determine if a date is a weekend
}
