% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{remove_weekends}
\alias{remove_weekends}
\title{Remove weekends from the data}
\usage{
remove_weekends(
  dt_data,
  weekend_days = c("Saturday", "Sunday"),
  abbreviate = FALSE
)
}
\arguments{
\item{dt_data}{object containing date data}

\item{weekend_days}{string indicating the days that are classified as weekend, e.g. "Saturday"}

\item{abbreviate}{logical indicating whether the character vector of weekend days is abbreviated or not}
}
\value{
data object with weekend days removed
}
\description{
Remove weekends from the data
}
