% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_annual.R
\name{to_annual.data.frame}
\alias{to_annual.data.frame}
\title{Convert object of class \code{data.frame} to annual periodicity}
\usage{
\method{to_annual}{data.frame}(dt_data, ...)
}
\arguments{
\item{dt_data}{data.frame with columns date and value}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{data.frame} to annual periodicity
}
