% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_annual.R
\name{to_annual.Date}
\alias{to_annual.Date}
\title{Convert object of class \code{Date} to annual periodicity}
\usage{
\method{to_annual}{Date}(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to annual frequency}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{Date} to annual periodicity
}
