% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_annual.R
\name{to_annual}
\alias{to_annual}
\title{Convert non-annual time series data to annual}
\usage{
to_annual(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to annual frequency}

\item{...}{arguments for other methods}
}
\description{
Convert non-annual time series data to annual
}
