% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_monthly.R
\name{to_monthly.data.frame}
\alias{to_monthly.data.frame}
\title{Convert object of class \code{data.frame} to monthly periodicity}
\usage{
\method{to_monthly}{data.frame}(dt_data, ...)
}
\arguments{
\item{dt_data}{data.frame with columns date and value}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{data.frame} to monthly periodicity
}
