% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_monthly.R
\name{to_monthly.Date}
\alias{to_monthly.Date}
\title{Convert object of class \code{Date} to monthly periodicity}
\usage{
\method{to_monthly}{Date}(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to monthly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{Date} to monthly periodicity
}
