% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_monthly.R
\name{to_monthly}
\alias{to_monthly}
\title{Convert non-monthly time series data to monthly}
\usage{
to_monthly(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to monthly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert non-monthly time series data to monthly
}
