% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_period.R
\name{to_period}
\alias{to_period}
\title{convert time series data in a \code{data.frame} to a different frequency}
\usage{
to_period(
  dt_data,
  period = c("weekly", "monthly", "quarterly", "annual"),
  aggregate_by = c("last", "sum", "avg", "median")
)
}
\arguments{
\item{dt_data}{date data to aggregate to monthly frequency}

\item{period}{string indicating the periodicity to convert data to}

\item{aggregate_by}{string indicating how to aggregate data}
}
\value{
\code{data.frame} or date vector depending on argument \code{dt_data}
}
\description{
convert time series data in a \code{data.frame} to a different frequency
}
