% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_quarterly.R
\name{to_quarterly.data.frame}
\alias{to_quarterly.data.frame}
\title{Convert object of class \code{data.frame} to quarterly periodicity}
\usage{
\method{to_quarterly}{data.frame}(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to quarterly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{data.frame} to quarterly periodicity
}
