% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_quarterly.R
\name{to_quarterly.Date}
\alias{to_quarterly.Date}
\title{Convert object of class \code{Date} to quarterly periodicity}
\usage{
\method{to_quarterly}{Date}(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to quarterly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert object of class \code{Date} to quarterly periodicity
}
