% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_quarterly.R
\name{to_quarterly}
\alias{to_quarterly}
\title{Convert non-quarterly time series data to quarterly}
\usage{
to_quarterly(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to quarterly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert non-quarterly time series data to quarterly
}
