% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_weekly.R
\name{to_weekly.data.frame}
\alias{to_weekly.data.frame}
\title{Convert to weekly data observations}
\usage{
\method{to_weekly}{data.frame}(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to weekly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert to weekly data observations
}
