% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_weekly.R
\name{to_weekly}
\alias{to_weekly}
\title{Convert to weekly observations}
\usage{
to_weekly(dt_data, ...)
}
\arguments{
\item{dt_data}{date data to aggregate to weekly frequency}

\item{...}{arguments for other methods}
}
\description{
Convert to weekly observations
}
