% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{get_date_col}
\alias{get_date_col}
\title{Find and rename the date column}
\usage{
get_date_col(dt_data)
}
\arguments{
\item{dt_data}{\code{data.frame} containing a column of class \code{Date}}
}
\value{
\code{data.frame} with column of class \code{Date} renamed as "date" or NULL if no date class col detected
}
\description{
Find and rename the date column
}
